restructured components/routes
