+++ src/router/search.vue
